move models into api folder
